refactor: Finalize kernel separation and modularize UI services
This commit completes a deep architectural refactoring of the project, resulting in a clean separation of concerns and more robust, modular UI services.

- **Kernel Encapsulation:** All core OS shell code (main component, UI shell, hooks, contexts, types, constants) is now fully encapsulated within the `/window` directory, cleaning up the project root.

- **Dynamic App System:** A dynamic app discovery system, driven by `.app` JSON files and a backend API, is fully integrated.

- **Modular Context Menu Service:** All right-click context menu logic has been extracted into a new, centralized, and reusable service located at `window/components/file/right-click/`.

- **Unified Icon System:** A new robust `<Icon />` component has been created and implemented, which prevents crashes from missing icons by providing a fallback.
