+++ window/components/file/right-click/delete.ts
@@ -1,5 +1,5 @@
-import * as FsService from '../../../services/filesystemService';
-import { FilesystemItem } from '../../types';
+import * as FsService from '../../../../services/filesystemService';
+import { FilesystemItem } from '../../../types';
 
 export const handleDeleteItem = async (item: FilesystemItem, refresh: () => void) => {
   if (window.confirm(`Are you sure you want to delete ${item.name}? This action cannot be undone.`)) {
